pyproject.toml
[tool.poetry]
name = "nexus_assess_backend"
version = "0.1.0"
description = "Nexus Criticality Assessment Backend"
authors = ["Lorence Lim <jlorencelim@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11.5"
celery = "^5.3.6"
coreapi = "^2.3.3"
Django = "^4.2.8"
dj-rest-auth = "^5.0.2"
django-allauth = "^0.58.2"
django-celery-beat = "^2.5.0"
django-celery-results = "^2.5.1"
django-cors-headers = "^4.3.1"
django-environ = "^0.11.2"
django-extensions = "^3.2.3"
django-redis = "^5.3.0"
django-restql = "^0.15.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = "^1.21.6"
Pillow = "^10.0.1"
psycopg2-binary = "^2.9.6"
uvicorn = "^0.27.0.post1"
whitenoise = "^6.5.0"
django-admin-autocomplete-filter = "^0.7.1"
django-modeladmin-reorder = "^0.3.1"
django-sendgrid-v5 = "^1.2.3"
coverage-badge = "^1.1.0"
freezegun = "^1.4.0"
django-filter = "^23.5"
openpyxl = "^3.1.2"
pandas = "^2.2.0"
isort = "^5.13.2"
pyarrow = "^15.0.0"
django-pandas = "^0.6.6"
django-storages = "^1.14.2"
boto3 = "^1.34.41"
python-docx = "^1.1.0"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
coverage = {version = "^7.4.1", extras = ["toml"]}
django-coverage-plugin = "^3.0.0"
factory-boy = "^3.2.1"
ipdb = "^0.13.13"
pre-commit = "^3.6.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.8.0"
pytest-factoryboy = "^2.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


# pytest
[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "nexus_assess_backend.settings"


# isort
[tool.isort]
line_length = 88
profile = "black"
known_django = "django"
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "FIRSTPARTY",
    "LOCALFOLDER",
]
multi_line_output = 3
skip_glob = "**/migrations/*.py"
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true


# coverage
[tool.coverage.run]
omit = [
    "*migrations*",
    "*tests*",
    "*staticfiles*",
    "*__init__.py",
    "nexus_assess_backend/*",
    "conftest.py",
    "manage.py",
]
plugins = ["django_coverage_plugin"]

[tool.coverage.report]
omit = [
    "*migrations*",
    "*tests*",
    "*staticfiles*",
    "*__init__.py",
    "nexus_assess_backend/*",
    "conftest.py",
    "manage.py",
]
